+++ lib/src/db/isoexpress/rt_system_demand_hourly.dart
@@ -12,27 +12,27 @@ import '../converters.dart';
 import 'package:elec_server/src/utils/iso_timestamp.dart';
 
 class RtSystemDemandReportArchive extends DailyIsoExpressReport {
-  @override
-  var location = getLocation('America/New_York');
+  // @override
+  // var location = getLocation('America/New_York');
 
   RtSystemDemandReportArchive({ComponentConfig? dbConfig, String? dir}) {
     dbConfig ??= ComponentConfig(
           host: '127.0.0.1', dbName: 'isoexpress', collectionName: 'system_demand');
     this.dbConfig = dbConfig;
-    dir ??= baseDir + 'EnergyReports/RtHourlyDemand/Raw/';
+    dir ??= '${baseDir}EnergyReports/RtHourlyDemand/Raw/';
     this.dir = dir;
+    reportName = 'Real-Time Hourly System Load Report';
   }
+  
+  
   @override
-  String reportName = 'Real-Time Hourly System Load Report';
+  String getUrl(Date asOfDate) =>
+      'https://www.iso-ne.com/transform/csv/hourlysystemdemand?start=${yyyymmdd(asOfDate)}&end=${yyyymmdd(asOfDate)}';
+
   @override
-  String getUrl(Date? asOfDate) =>
-      'https://www.iso-ne.com/transform/csv/hourlysystemdemand?start=' +
-      yyyymmdd(asOfDate) +
-      '&end=' +
-      yyyymmdd(asOfDate);
-  @override
-  File getFilename(Date? asOfDate) =>
-      File(dir + 'rt_hourlydemand_' + yyyymmdd(asOfDate) + '.csv');
+  File getFilename(Date asOfDate) =>
+      File('${dir}rt_hourlydemand_${yyyymmdd(asOfDate)}.csv');
+
 
   /// File may be incomplete if downloaded during the day ...
   @override
@@ -52,10 +52,6 @@ class RtSystemDemandReportArchive extends DailyIsoExpressReport {
     row['hourBeginning'] = [];
     row['Total Load'] = <num>[];
     for (var e in rows) {
-//      var hb = parseHourEndingStamp(localDate, e['Hour Ending']);
-//      row['hourBeginning'].add(TZDateTime.fromMillisecondsSinceEpoch(
-//              location, hb.millisecondsSinceEpoch)
-//          .toIso8601String());
       row['hourBeginning'].add(parseHourEndingStamp(localDate, e['Hour Ending']));
       row['Total Load'].add(e['Total Load']);
     }
@@ -67,14 +63,6 @@ class RtSystemDemandReportArchive extends DailyIsoExpressReport {
     var data = mis.readReportTabAsMap(file, tab: 0);
     if (data.isEmpty) return <Map<String, dynamic>>[];
     return [converter(data)];
-  }
-
-  /// Check if this date is in the db already
-  Future<bool> hasDay(Date date) async {
-    var res =
-        await dbConfig.coll.findOne({'market': 'RT', 'date': date.toString()});
-    if (res == null || res.isEmpty) return false;
-    return true;
   }
 
   /// Recreate the collection from scratch.
@@ -89,22 +77,5 @@ class RtSystemDemandReportArchive extends DailyIsoExpressReport {
     await dbConfig.db.createIndex(dbConfig.collectionName,
         keys: {'market': 1, 'date': 1}, unique: true);
     await dbConfig.db.close();
-  }
-
-  Future<Map<String, String?>> lastDay() async {
-    List pipeline = [];
-    pipeline.add({
-      '\$group': {
-        '_id': 0,
-        'lastDay': {'\$max': '\$date'}
-      }
-    });
-    Map res = await dbConfig.coll.aggregate(pipeline);
-    return {'lastDay': res['result'][0]['lastDay']};
-  }
-
-  Date lastDayAvailable() => Date.today(location: UTC).subtract(2);
-  Future<void> deleteDay(Date day) async {
-    return await (dbConfig.coll.remove(mongo.where.eq('date', day.toString())) as FutureOr<void>);
   }
 }