+++ lib/src/db/isone/variable_energy_resource.dart
@@ -1,0 +1,7 @@
+library db.isone.variable_energy_resource;
+
+/// Great dataset
+/// https://www.iso-ne.com/system-planning/planning-models-and-data/variable-energy-resource-data
+/// saved in Archive/Isone/VariableEnergySource/Raw
+///
+